move more stuff to avoid circular imports
